requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
cffi==2.0.0
cryptography==47.0.0
Django==6.0.4
django-ninja==1.6.2
django-silk==5.5.0
ecdsa==0.19.2
gprof2dot==2025.4.14
passlib==1.7.4
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
